feat: [005-SEARCH-STORE]
별점 순 상점 검색 기능

+++ src/main/java/com/project/reservation/model/response/StoreResponse.java
@@ -34,7 +34,7 @@ public class StoreResponse {
         private final String storeName;
         private final String description;
 
-        private final Integer star;
+        private final Double star;
         private final Boolean isAvail;
 
         private final Address address;
